requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.2.0
